demo: emit BreadcrumbList JSON-LD via Breadcrumbs component
Splits Breadcrumbs into a server component that emits the BreadcrumbList JSON-LD using the new `JsonLd` component, and a client component (`BreadcrumbsClient`) that contains the existing interactive UI verbatim. Adds `domain` to the scope fragment so the server component can resolve the absolute URL via `getSiteConfigForDomain`.
