+++ demo/site/src/common/components/breadcrumbs/Breadcrumbs.fragment.ts
@@ -9,6 +9,7 @@ export const breadcrumbsFragment = gql`
             path
         }
         scope {
+            domain
             language
         }
     }
